+++ plugin-api/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @matatbread/matbot-plugin-api
 
+## 0.2.9
+
 ## 0.2.8
 
 ### Patch Changes