+++ plugins/cognition/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @matatbread/matbot-cognition
 
+## 0.2.9
+
+### Patch Changes
+
+- @matatbread/matbot-plugin-api@0.2.9
+- @matatbread/matbot-tool-store@0.2.9
+
 ## 0.2.8
 
 ### Patch Changes