+++ plugins/storage/sqlite/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @matatbread/matbot-storage-sqlite
 
+## 0.2.9
+
+### Patch Changes
+
+- Updated dependencies
+  - @matatbread/matbot-core@0.2.9
+  - @matatbread/matbot-plugin-api@0.2.9
+
 ## 0.2.8
 
 ### Patch Changes